[Suggestion] Use Phosh instead of Lightdm (#85)
I know there was a time where phosh was not working as intended, but now that it can be used with phosh.service, I'm proposing here to replace lightdm-mobile-greeter.

The only reason I can remember to keep lightdm was for the toggle to switch to GNOME. But I think it is easy enough to set up a different manager (such as gdm even) that it isn't really necessary for a base image, and this will reduce another package required. 

Further, phosh already switches to using itself as a lock screen, so lightdm looks a bit out of place initially. But that is just aesthetics. 

I also did systemctl disable gdm, since for whatever reason on my image I made with this, it seems gdm tries to init before phosh takes over.